clarify(weight-at-age): fleet = -2 is fecundity
* be specific in that fleet -2 is equal to fecundity, where the previous text
was slightly unclear where it said fecundity*maturity*weight-at-age, I changed
this to fecundity-scalar after discussing with Aaron Berger so that users know
that this matrix that corresponds to fleet -2 should equal fecundity but doesn't
tell people exactly how to calculate it. The example is what is done for hake,
where fecundity-scalar = 1.
* change wt-at-age to weight-at-age
* fix two typos; random N at the end of a bullet and age 0 fish to age-0 fish