Comment out lines that are causing it to fail
I am not sure if these lines actually have errors in them but if you
comment them out then it runs. I am not getting the same errors if I
turn off the tagging. I think there must be something else going on in
the document that is causing the failure because it is not actually giving
me a warning about these lines directly. Wrapping them in \texttt did not
work either.